# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/goods/": {
            "get": {
                "description": "Returns list of goods",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Get all goods",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/goods/create": {
            "post": {
                "description": "Creates a goods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Create good",
                "parameters": [
                    {
                        "description": "Good details",
                        "name": "good",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Good"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/goods/delete/": {
            "post": {
                "description": "Deletes goods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Delete goods",
                "parameters": [
                    {
                        "description": "Good IDs",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.MultiID"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/goods/delete/{id}": {
            "delete": {
                "description": "Deletes a goods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Delete a good",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Good ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/goods/show/{id}": {
            "get": {
                "description": "Returns one good",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Get one good",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Good ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/goods/update/{id}": {
            "put": {
                "description": "Updates a goods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goods"
                ],
                "summary": "Update a good",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Good ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Good details",
                        "name": "good",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Good"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.MultiID": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dtos.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Feature": {
            "type": "object",
            "properties": {
                "feature": {
                    "type": "string"
                },
                "featureID": {
                    "type": "integer"
                },
                "isDeleted": {
                    "type": "boolean"
                }
            }
        },
        "models.Good": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "feature": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Feature"
                    }
                },
                "goodID": {
                    "type": "integer"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "vendor": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vendor"
                    }
                }
            }
        },
...
        "models.Vendor": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "inn": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "ogrn": {
                    "type": "string"
                },
                "orgName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "vendorID": {
                    "type": "integer"
                }
            }
        }
    }
}